Add import and export feature
- Add ability to import flashcards
- Add ability to export flashcards
    - Reformat script.js
- CSS tweak to adjust positioning for flashcard deletion button

+++ script.js
@@ -1,3 +1,6 @@
+
+let container = localStorage.getItem("items") ? JSON.parse(localStorage.getItem("items")) : [];
+
 buildFlashcard = (element, index) => {
   const flashcard = document.createElement("article");
   const wrapper = document.createElement("section");
@@ -33,7 +36,7 @@ buildFlashcard = (element, index) => {
   flashcard.appendChild(wrapper);
   
   document.querySelector("main").appendChild(flashcard);
-}
+};
 
 createFlashcard = () => {
   const question = document.querySelector("#app-form-question");
@@ -42,45 +45,75 @@ createFlashcard = () => {
   let flashcard_info = {
     'question': question.value,
     'answer': answer.value
-  }
+  };
 
   container.push(flashcard_info);
   localStorage.setItem('items', JSON.stringify(container));
   buildFlashcard(container[container.length - 1], container.length - 1);
   question.value = "";
   answer.value = "";
-}
-
-let container = localStorage.getItem("items") ? JSON.parse(localStorage.getItem("items")) : [];
+};
 
 container.forEach(buildFlashcard);
 
 // display form
-document.
-  querySelector("#app-form-show").
-  addEventListener("click", () => {
-    document.querySelector("#app-form").style.display = "flex";
-  });
+document.querySelector("#app-form-show").addEventListener("click", () => {
+  document.querySelector("#app-form").style.display = "flex";
+});
 
 // hide form
-document.
-  querySelector("#app-form-hide").
-  addEventListener("click", () => {
-    document.querySelector("#app-form").style.display = "none";
-  });
+document.querySelector("#app-form-hide").addEventListener("click", () => {
+  document.querySelector("#app-form").style.display = "none";
+});
 
 // save form contents to local storage
-document.
-  querySelector("#app-form-save").
-  addEventListener("click", () => {
-    createFlashcard();
-  });
+document.querySelector("#app-form-save").addEventListener("click", () => {
+  createFlashcard();
+});
 
 // delete flashcards
-document.
-  querySelector("#app-main-reset").
-  addEventListener("click", () => {
+document.querySelector("#app-main-reset").addEventListener("click", () => {
+  localStorage.clear();
+  document.querySelector("main").innerHTML = '';
+  container = [];
+});
+
+// Export Storage object to JSON file
+document.querySelector("#app-main-export").addEventListener("click", () => {
+  let dataStr = JSON.stringify(container);
+  let dataUri = 'data:application/json;charset=utf-8,'+ encodeURIComponent(dataStr);
+
+  let exportFileDefaultName = `${crypto.randomUUID()}.json`;
+  
+  let linkElement = document.createElement('a');
+  
+  linkElement.setAttribute('href', dataUri);
+  linkElement.setAttribute('download', exportFileDefaultName);
+  linkElement.click();
+});
+
+// Import Storage object from JSON file
+document.querySelector("#app-main-import").addEventListener("click", () => {
+  document.querySelector("#app-main-import-json").click();
+});
+
+document.querySelector("#app-main-import-json").addEventListener("change", (e) => {
+  const file = e.target.files[0];
+
+  if (null == file) {
+    console.log("oops! something went really wrong.")
+    return;
+  }
+
+  file.text().then((context) => {
+    let counter = 0;
     localStorage.clear();
     document.querySelector("main").innerHTML = '';
-    container = [];
+    localStorage.setItem('items', context);
+    container = JSON.parse(context);
+    for (let flashcard of container) {
+      buildFlashcard(flashcard, counter);
+      counter += 1;
+    }
   });
+});
